Move Janus classes to separate files, no functional changes.
