+++ publisher_stats_counter_test.go
